Rename project to 'quadcopter'
'flight_controller_v2' was the code name for the project based
on the fact that there was a development version 'v1' of the PCB
designed in Eagle which was not really mature. Except for the idea
this is a completely new project - KiCAD, different components etc.
Use generic name 'quadcopter' for now until the projects grows
to an actual product.