Add finished PCB design for initial release
This is quite a messy version of the initial PCB, hopefully without
errors. Should conform to JLCPCB manufacturing process.
This is a 4 layer PCB with components on the TOP, test pads and
placement for debug connector on the bottom.
The inside layers consist of a ground plane and a layer for power
signals, especially 3v3 and BATT.